Extrator_mysql/requirements.txt
sqlalchemy
pandas~=2.2.3
pyarrow~=19.0.0
python-dotenv~=1.0.1
boto3~=1.36.2
chardet~=5.2.0
gitpython
pymongo~=4.10.1
requests>=2.28.1
urllib3>=1.26.14
charset-normalizer>=2.0.12
azure-storage-blob
azure-core~=1.32.0
pymysql
colorlog
pytz~=2024.2
polars >=0.19.0
deltalake >=0.14.0
rich
loguru
structlog
humanfriendly
psutil
fastparquet
polars-lts-cpu #Para maior compatibilidade e estabilidade
mysql
mysql-connector-python
